Generalized blocked PCA to support projection of original expression.
This combines the functionality of the old MultiBatchPca and ResidualPca
classes. To reproduce the former, we compute the rotation vectors from the
residuals but project the original expression values, so as to retain the
benefits of blocking without making assumptions about the block differences.

We rename the residual_pca namespace to blocked_pca to reflects its generality.
Because we don't need to support separate MultiBatchPca functionality, we can
consolidate the utilities into the blocked_pca namespace and get rid of some
pca_utils files. The remaining utilities can be squashed into a single file.

Also reduced the number of PCs to extract in the tests, given that the
differences in numerical precision start to exceed tolerances at later PCs.